+++ activesupport/lib/active_support/core_ext/date/zones.rb
@@ -2,21 +2,6 @@ require 'date'
 require 'active_support/core_ext/time/zones'
 
 class Date
-  # *DEPRECATED*: Use +Date#in_time_zone+ instead.
-  #
-  # Converts Date to a TimeWithZone in the current zone if <tt>Time.zone</tt> or
-  # <tt>Time.zone_default</tt> is set, otherwise converts Date to a Time via
-  # Date#to_time.
-  def to_time_in_current_zone
-    ActiveSupport::Deprecation.warn 'Date#to_time_in_current_zone is deprecated. Use Date#in_time_zone instead', caller
-
-    if ::Time.zone
-      ::Time.zone.local(year, month, day)
-    else
-      to_time
-    end
-  end
-
   # Converts Date to a TimeWithZone in the current zone if Time.zone or Time.zone_default
   # is set, otherwise converts Date to a Time via Date#to_time
   #